requirements.txt
torch>=1.12.0
torchvision>=0.13.0
pytorch-lightning>=1.8.0
timm>=0.6.0
pandas>=1.4.0
numpy>=1.21.0
scikit-learn>=1.1.0
scikit-image>=0.19.0
Pillow>=9.0.0
PyYAML>=6.0
tqdm>=4.64.0
tensorboard>=2.8.0
